requirements.txt
psycopg2-binary
faker
pandas
ipykernel
pyarrow
dask
